feat: moved earn feature flag selectors into Earn directory for code ownership
