+++ app/components/UI/Earn/Views/EarnWithdrawInputView/EarnWithdrawInputView.test.tsx
@@ -132,12 +132,9 @@ jest.mock('../../../Stake/hooks/usePoolStakedUnstake', () => ({
 
 jest.mock('../../../../../selectors/featureFlagController/confirmations');
 
-jest.mock(
-  '../../../../../selectors/featureFlagController/earnFeatureFlags',
-  () => ({
-    selectStablecoinLendingEnabledFlag: jest.fn().mockReturnValue(false),
-  }),
-);
+jest.mock('../../selectors/featureFlags', () => ({
+  selectStablecoinLendingEnabledFlag: jest.fn().mockReturnValue(false),
+}));
 
 describe('UnstakeInputView', () => {
   const selectConfirmationRedesignFlagsMock = jest.mocked(selectConfirmationRedesignFlags)
