+++ app/components/UI/Earn/hooks/useEarnTokenDetails.ts
@@ -16,7 +16,7 @@ import { addCurrencySymbol } from '../../../../util/number';
 import useBalance from '../../Stake/hooks/useBalance';
 import { TokenI } from '../../Tokens/types';
 import { deriveBalanceFromAssetMarketDetails } from '../../Tokens/util';
-import { selectStablecoinLendingEnabledFlag } from '../../../../selectors/featureFlagController/earnFeatureFlags';
+import { selectStablecoinLendingEnabledFlag } from '../selectors/featureFlags';
 
 // Mock APR values - will be replaced with real API data later
 const MOCK_APR_VALUES: { [symbol: string]: string } = {
